services/web/requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
psycopg2-binary==2.8.4

absl-py==0.11.0
astunparse==1.6.3
blinker==1.4
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
curl==0.0.1
dnspython==2.0.0
email-validator==1.1.2

Flask-Login==0.5.0
Flask-Mail==0.9.1

Flask-WTF==0.14.3
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2

Markdown==3.3.3
MarkupSafe==1.1.1

oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.15.0
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6

six==1.15.0
SQLAlchemy==1.3.22

termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
